Reuse erc20 actions
