+++ application/comit_node/src/swap_protocols/rfc003/bob/actions/erc20.rs
@@ -2,74 +2,19 @@ use crate::swap_protocols::{
     asset::Asset,
     ledger::Ethereum,
     rfc003::{
-        actions::{non_erc20::CreateActions, Actions},
+        actions::{erc20, non_erc20::CreateActions, Actions},
         bob::{
             self,
             actions::{Accept, Decline},
             SwapCommunication,
         },
-        ethereum::{self, Erc20Htlc},
+        ethereum,
         state_machine::HtlcParams,
-        Action, Ledger, LedgerState, Secret,
+        Action, Ledger, LedgerState,
     },
 };
-use ethereum_support::{Bytes, Erc20Token, EtherQuantity, Network};
+use ethereum_support::Erc20Token;
 use std::sync::Arc;
-
-fn deploy_action(htlc_params: HtlcParams<Ethereum, Erc20Token>) -> ethereum::ContractDeploy {
-    htlc_params.into()
-}
-
-pub fn fund_action(
-    htlc_params: HtlcParams<Ethereum, Erc20Token>,
-    to_erc20_contract: ethereum_support::Address,
-    beta_htlc_location: ethereum_support::Address,
-) -> ethereum::SendTransaction {
-    let htlc = Erc20Htlc::from(htlc_params.clone());
-    let gas_limit = Erc20Htlc::fund_tx_gas_limit();
-    let network = htlc_params.ledger.network;
-
-    ethereum::SendTransaction {
-        to: to_erc20_contract,
-        data: htlc.funding_tx_payload(beta_htlc_location),
-        gas_limit,
-        amount: EtherQuantity::zero(),
-        network,
-    }
-}
-
-pub fn refund_action(
-    network: Network,
-    beta_htlc_location: ethereum_support::Address,
-) -> ethereum::SendTransaction {
-    let data = Bytes::default();
-    let gas_limit = Erc20Htlc::tx_gas_limit();
-
-    ethereum::SendTransaction {
-        to: beta_htlc_location,
-        data,
-        gas_limit,
-        amount: EtherQuantity::zero(),
-        network,
-    }
-}
-
-fn redeem_action(
-    alpha_htlc_location: ethereum_support::Address,
-    secret: Secret,
-    network: Network,
-) -> ethereum::SendTransaction {
-    let data = Bytes::from(secret.raw_secret().to_vec());
-    let gas_limit = Erc20Htlc::tx_gas_limit();
-
-    ethereum::SendTransaction {
-        to: alpha_htlc_location,
-        data,
-        gas_limit,
-        amount: EtherQuantity::zero(),
-        network,
-    }
-}
 
 impl<AL, AA> Actions for bob::State<AL, Ethereum, AA, Erc20Token>
 where
@@ -113,40 +58,40 @@ where
 
         use self::LedgerState::*;
 
-        let mut actions = match (alpha_state, beta_state, self.secret) {
-            (Funded { htlc_location, .. }, _, Some(secret)) => {
-                vec![bob::ActionKind::Redeem(<(AL, AA)>::redeem_action(
-                    HtlcParams::new_alpha_params(request, response),
-                    htlc_location.clone(),
-                    &*self.secret_source,
-                    secret,
-                ))
-                .into_action()]
-            }
-            (Funded { .. }, NotDeployed, _) => {
-                vec![
-                    bob::ActionKind::Deploy(deploy_action(HtlcParams::new_beta_params(
-                        request, response,
-                    )))
-                    .into_action(),
-                ]
-            }
-            (Funded { .. }, Deployed { htlc_location, .. }, _) => {
-                vec![bob::ActionKind::Fund(fund_action(
-                    HtlcParams::new_beta_params(request, response),
-                    request.beta_asset.token_contract,
-                    *htlc_location,
-                ))
-                .into_action()]
-            }
-            _ => vec![],
-        };
+        let mut actions =
+            match (alpha_state, beta_state, self.secret) {
+                (Funded { htlc_location, .. }, _, Some(secret)) => {
+                    vec![bob::ActionKind::Redeem(<(AL, AA)>::redeem_action(
+                        HtlcParams::new_alpha_params(request, response),
+                        htlc_location.clone(),
+                        &*self.secret_source,
+                        secret,
+                    ))
+                    .into_action()]
+                }
+                (Funded { .. }, NotDeployed, _) => vec![bob::ActionKind::Deploy(
+                    erc20::deploy_action(HtlcParams::new_beta_params(request, response)),
+                )
+                .into_action()],
+                (Funded { .. }, Deployed { htlc_location, .. }, _) => {
+                    vec![bob::ActionKind::Fund(erc20::fund_action(
+                        HtlcParams::new_beta_params(request, response),
+                        request.beta_asset.token_contract,
+                        *htlc_location,
+                    ))
+                    .into_action()]
+                }
+                _ => vec![],
+            };
 
         if let Funded { htlc_location, .. } = beta_state {
             actions.push(
-                bob::ActionKind::Refund(refund_action(request.beta_ledger.network, *htlc_location))
-                    .into_action()
-                    .with_invalid_until(request.beta_expiry),
+                bob::ActionKind::Refund(erc20::refund_action(
+                    request.beta_ledger.network,
+                    *htlc_location,
+                ))
+                .into_action()
+                .with_invalid_until(request.beta_expiry),
             );
         }
         actions
@@ -197,7 +142,7 @@ where
         let mut actions =
             match (alpha_state, beta_state, self.secret) {
                 (Funded { htlc_location, .. }, _, Some(secret)) => vec![bob::ActionKind::Redeem(
-                    redeem_action(*htlc_location, secret, request.alpha_ledger.network),
+                    erc20::redeem_action(*htlc_location, secret, request.alpha_ledger.network),
                 )
                 .into_action()],
                 (Funded { .. }, NotDeployed, _) => vec![bob::ActionKind::Fund(
